% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom-lvplot.r
\name{geom_lvplot}
\alias{geom_lvplot}
\title{Letter value plot.}
\usage{
geom_lvplot(mapping = NULL, data = NULL, stat = "lvplot",
  position = "dodge", outlier.colour = "black", outlier.shape = 19,
  outlier.size = 1.5, outlier.stroke = 0.5, varwidth = FALSE,
  show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), is combined with the default mapping at the top level of the
plot. You only need to supply \code{mapping} if there isn't a mapping
defined for the plot.}

\item{data}{A data frame. If specified, overrides the default data frame
defined at the top level of the plot.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{outlier.colour}{Override aesthetics used for the outliers. Defaults
come from \code{geom_point()}.}

\item{outlier.shape}{Override aesthetics used for the outliers. Defaults
come from \code{geom_point()}.}

\item{outlier.size}{Override aesthetics used for the outliers. Defaults
come from \code{geom_point()}.}

\item{outlier.stroke}{Override aesthetics used for the outliers. Defaults
come from \code{geom_point()}.}

\item{varwidth}{if \code{FALSE} (default) make a standard box plot. If
\code{TRUE}, boxes are drawn with widths proportional to the
square-roots of the number of observations in the groups (possibly
weighted, using the \code{weight} aesthetic).}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{...}{other arguments passed on to \code{\link{layer}}. There are
  three types of arguments you can use here:

  \itemize{
  \item Aesthetics: to set an aesthetic to a fixed value, like
     \code{color = "red"} or \code{size = 3}.
  \item Other arguments to the layer, for example you override the
    default \code{stat} associated with the layer.
  \item Other arguments passed on to the stat.
  }}

\item{geom,stat}{Use to override the default connection between
\code{geom_lvplot} and \code{stat_lvplot}.}
}
\description{
See McGill et al. (1978) for more details.
}
\section{Aesthetics}{

}
\examples{
library(ggplot2)
p <- ggplot(mpg, aes(class, hwy))
p + geom_lvplot()
p + geom_lvplot() + geom_jitter(width = 0.2)
p + geom_lvplot() + coord_flip()
p + geom_lvplot(alpha=1, aes(fill=..LV..)) + scale_fill_brewer()

p + geom_lvplot(varwidth = TRUE)
p + geom_lvplot(fill = "white", colour = "#3366FF")
p + geom_lvplot(outlier.colour = "red", outlier.shape = 1)

# Boxplots are automatically dodged when any aesthetic is a factor
p + geom_lvplot(aes(fill = drv))

# You can also use boxplots with continuous x, as long as you supply
# a grouping variable. cut_width is particularly useful
ggplot(diamonds, aes(carat, price)) +
  geom_lvplot()
ggplot(diamonds, aes(carat, price)) +
  geom_lvplot(aes(group = cut_width(carat, 0.25)))
}
\references{
McGill, R., Tukey, J. W. and Larsen, W. A. (1978) Variations of
    box plots. The American Statistician 32, 12-16.
}
\seealso{
\code{\link{stat_quantile}} to view quantiles conditioned on a
  continuous variable.
}

