% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stat-lvplot.r
\name{lvtable}
\alias{lvtable}
\title{Compute table of k letter values for vector x}
\usage{
lvtable(x, k, alpha = 0.95)
}
\arguments{
\item{x}{input numeric vector}

\item{k}{number of letter values to compute}

\item{alpha}{alpha-threshold for confidence level}
}
\value{
data frame of letter value (upper and lower value, if not the median), depth, and confidence interval.
}
\description{
Calculate k letter values from numeric vector x together with an 100(1-alpha)\% confidence interval.
}

