% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stat-lvplot.r
\name{stat_lvplot}
\alias{stat_lvplot}
\title{title of stat_lvplot}
\usage{
stat_lvplot(mapping = NULL, data = NULL, geom = "lvplot",
  position = "dodge", na.rm = FALSE, coef = 1.5, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link{aes}} or \code{\link{aes_string}}. Only needs to be set
at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override
the plot defaults.}

\item{geom}{The geometric object to use display the data}

\item{position}{The position adjustment to use for overlapping points
on this layer}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with
a warning.  If \code{TRUE} silently removes missing values.}

\item{coef}{length of the whiskers as multiple of IQR.  Defaults to 1.5}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{...}{other arguments passed on to \code{\link{layer}}. This can
include aesthetics whose values you want to set, not map. See
\code{\link{layer}} for more details.}
}
\description{
need a better description here, probably referring to geom_lvplot
}
\section{Computed variables}{

\describe{
  \item{width}{width of boxplot}
  \item{ymin}{lower whisker = smallest observation greater than or equal to lower hinge - 1.5 * IQR}
  \item{lower}{lower hinge, 25\% quantile}
  \item{middle}{median, 50\% quantile}
  \item{upper}{upper hinge, 75\% quantile}
  \item{ymax}{upper whisker = largest observation less than or equal to upper hinge + 1.5 * IQR}
}
}

